moved tests next to files
