+++ resources/app/components/NoColorsFound.test.tsx
@@ -1,10 +1,10 @@
 import React from 'react';
 import { render, cleanup, fireEvent } from 'react-testing-library';
-import NoColorsFound from '../NoColorsFound';
-import { closeWindow } from '../../helpers/window'; 
+import NoColorsFound from './NoColorsFound';
+import { closeWindow } from '../helpers/window'; 
 
-jest.mock('../../assets/minimilistBubbles.svg', () => 'bubbles-img');
-jest.mock('../../helpers/window', () => {
+jest.mock('../assets/minimilistBubbles.svg', () => 'bubbles-img');
+jest.mock('../helpers/window', () => {
 
   return {
     closeWindow: jest.fn(),
